회원 API에 대한 Type 정의 통일성있게 구현 필요 : refactor : FcmTokenCommand, FcmTokenDto 를 AuthCommand , AuthDto로 통함 및 DOCS 업데이트 https://github.com/Sejong-Balsamic/Malsami-FE/issues/455
