+++ src/main/java/com/balsamic/sejongmalsami/object/AuthCommand.java
@@ -1,5 +1,6 @@
 package com.balsamic.sejongmalsami.object;
 
+import com.balsamic.sejongmalsami.object.postgres.Member;
 import java.util.UUID;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
@@ -14,4 +15,6 @@ public class AuthCommand {
   private String refreshToken;
   private String accessToken;
   private UUID memberId;
+  private Member member;
+  private String fcmToken;
 }
