+++ src/main/java/com/balsamic/sejongmalsami/service/AuthService.java
@@ -3,7 +3,6 @@ package com.balsamic.sejongmalsami.service;
 import com.balsamic.sejongmalsami.object.AuthCommand;
 import com.balsamic.sejongmalsami.object.AuthDto;
 import com.balsamic.sejongmalsami.object.CustomUserDetails;
-import com.balsamic.sejongmalsami.object.FcmTokenCommand;
 import com.balsamic.sejongmalsami.object.MemberCommand;
 import com.balsamic.sejongmalsami.object.MemberDto;
 import com.balsamic.sejongmalsami.object.WebLoginDto;
@@ -119,7 +118,7 @@ public class AuthService {
    * 로그아웃 처리
    */
   @Transactional
-  public void logout(FcmTokenCommand command, HttpServletRequest request, HttpServletResponse response) {
+  public void logout(AuthCommand command, HttpServletRequest request, HttpServletResponse response) {
     // FCM 토큰 삭제
     fcmTokenService.deleteFcmToken(command);
