+++ src/main/java/com/balsamic/sejongmalsami/util/deprecated/FcmTokenCommand.java
@@ -1,4 +1,4 @@
-package com.balsamic.sejongmalsami.object;
+package com.balsamic.sejongmalsami.util.deprecated;
 
 import com.balsamic.sejongmalsami.object.postgres.Member;
 import lombok.AllArgsConstructor;
@@ -7,6 +7,7 @@ import lombok.Getter;
 import lombok.Setter;
 import lombok.ToString;
 
+@Deprecated
 @ToString
 @AllArgsConstructor
 @Getter
