+++ src/main/java/com/balsamic/sejongmalsami/util/deprecated/FcmTokenDto.java
@@ -1,10 +1,11 @@
-package com.balsamic.sejongmalsami.object;
+package com.balsamic.sejongmalsami.util.deprecated;
 
 import com.balsamic.sejongmalsami.object.mongo.FcmToken;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.ToString;
 
+@Deprecated
 @Builder
 @Getter
 @ToString
